+++ docs/.vuepress/config.js
@@ -104,6 +104,7 @@ module.exports = {
                 '/docs/v4/image/',
                 '/docs/v4/image-glide/',
                 '/docs/v4/image-picasso/',
+                '/docs/v4/inline-parser/',
                 '/docs/v4/linkify/',
                 '/docs/v4/recycler/',
                 '/docs/v4/recycler-table/',