+++ markwon-inline-parser/src/main/java/io/noties/markwon/inlineparser/AsteriskDelimiterProcessor.java
@@ -1,7 +1,0 @@
-package io.noties.markwon.inlineparser;
-
-/**
- * @since 4.2.0-SNAPSHOT
- */
-public class AsteriskDelimiterProcessor extends org.commonmark.internal.inline.AsteriskDelimiterProcessor {
-}