+++ sample/src/main/java/io/noties/markwon/sample/MainActivity.java
@@ -24,6 +24,7 @@ import io.noties.markwon.sample.customextension.CustomExtensionActivity;
 import io.noties.markwon.sample.customextension2.CustomExtensionActivity2;
 import io.noties.markwon.sample.editor.EditorActivity;
 import io.noties.markwon.sample.html.HtmlActivity;
+import io.noties.markwon.sample.inlineparser.InlineParserActivity;
 import io.noties.markwon.sample.latex.LatexActivity;
 import io.noties.markwon.sample.precomputed.PrecomputedActivity;
 import io.noties.markwon.sample.recycler.RecyclerActivity;
@@ -120,6 +121,10 @@ public class MainActivity extends Activity {
 
             case EDITOR:
                 activity = EditorActivity.class;
+                break;
+
+            case INLINE_PARSER:
+                activity = InlineParserActivity.class;
                 break;
 
             default: